requirements.txt

# Core dependencies
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
websockets==12.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
